jobmate/jobmate/pyproject.toml
[project]
name = "jobmate"
version = "0.1.0"
description = "AI Job Assistant using Gemini (Google Generative AI)"
readme = "README.md"
authors = [
    { name = "musman4it", email = "musman4it@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.3.2",  # Use the latest or stable version
    "python-dotenv>=1.1.0",
]

[project.scripts]
jobmate = "jobmate:main"
jobagent = "jobmate.jobmate:main"

[build-system]
requires = ["hatchling"]


build-backend = "hatchling.build"
